+++ Source/CesiumRuntime/Private/CesiumGltfComponent.cpp
@@ -1558,13 +1558,26 @@ static void SetGltfParameterValues(
         static_cast<float>(textureCoordinateSet.second));
   }
 
-  if (pbr.baseColorFactor.size() >= 3) {
+  if (pbr.baseColorFactor.size() > 3) {
     pMaterial->SetVectorParameterValueByInfo(
         FMaterialParameterInfo("baseColorFactor", association, index),
-        FVector(
+        FLinearColor(
             pbr.baseColorFactor[0],
             pbr.baseColorFactor[1],
-            pbr.baseColorFactor[2]));
+            pbr.baseColorFactor[2],
+            pbr.baseColorFactor[3]));
+  } else if (pbr.baseColorFactor.size() == 3) {
+    pMaterial->SetVectorParameterValueByInfo(
+        FMaterialParameterInfo("baseColorFactor", association, index),
+        FLinearColor(
+            pbr.baseColorFactor[0],
+            pbr.baseColorFactor[1],
+            pbr.baseColorFactor[2],
+            1.));
+  } else {
+    pMaterial->SetVectorParameterValueByInfo(
+        FMaterialParameterInfo("baseColorFactor", association, index),
+        FLinearColor(1., 1., 1., 1.));
   }
   pMaterial->SetScalarParameterValueByInfo(
       FMaterialParameterInfo("metallicFactor", association, index),
@@ -1849,14 +1862,26 @@ static void loadPrimitiveGameThreadPart(
   const FName ImportedSlotName(
       *(TEXT("CesiumMaterial") + FString::FromInt(nextMaterialId++)));
 
+  const auto is_in_blend_mode = [](auto& result) {
+    return !!result.pMaterial && result.pMaterial->alphaMode ==
+                                     CesiumGltf::Material::AlphaMode::BLEND;
+  };
+
 #if PLATFORM_MAC
   // TODO: figure out why water material crashes mac
-  UMaterialInterface* pBaseMaterial = pGltf->BaseMaterial;
+  UMaterialInterface* pBaseMaterial =
+      (is_in_blend_mode(loadResult) && pbr.baseColorFactor.size() > 3 &&
+       pbr.baseColorFactor[3] < 0.996) // 1. - 1. / 256.
+          ? pGltf->BaseMaterialWithTranslucency
+          : pGltf->BaseMaterial;
 #else
   UMaterialInterface* pBaseMaterial =
       (loadResult.onlyWater || !loadResult.onlyLand)
           ? pGltf->BaseMaterialWithWater
-          : pGltf->BaseMaterial;
+          : (is_in_blend_mode(loadResult) && pbr.baseColorFactor.size() > 3 &&
+             pbr.baseColorFactor[3] < 0.996) // 1. - 1. / 256.
+                ? pGltf->BaseMaterialWithTranslucency
+                : pGltf->BaseMaterial;
 #endif
 
   UMaterialInstanceDynamic* pMaterial = UMaterialInstanceDynamic::Create(
@@ -2015,6 +2040,7 @@ UCesiumGltfComponent::CreateOffGameThread(
     TUniquePtr<HalfConstructed> pHalfConstructed,
     const glm::dmat4x4& cesiumToUnrealTransform,
     UMaterialInterface* pBaseMaterial,
+    UMaterialInterface* pBaseTranslucentMaterial,
     UMaterialInterface* pBaseWaterMaterial,
     FCustomDepthParameters CustomDepthParameters,
     const Cesium3DTilesSelection::BoundingVolume& boundingVolume) {
@@ -2037,6 +2063,10 @@ UCesiumGltfComponent::CreateOffGameThread(
 
   if (pBaseMaterial) {
     Gltf->BaseMaterial = pBaseMaterial;
+  }
+
+  if (pBaseTranslucentMaterial) {
+    Gltf->BaseMaterialWithTranslucency = pBaseTranslucentMaterial;
   }
 
   if (pBaseWaterMaterial) {
@@ -2067,11 +2097,15 @@ UCesiumGltfComponent::UCesiumGltfComponent() : USceneComponent() {
   // Structure to hold one-time initialization
   struct FConstructorStatics {
     ConstructorHelpers::FObjectFinder<UMaterialInstance> BaseMaterial;
+    ConstructorHelpers::FObjectFinder<UMaterialInstance>
+        BaseMaterialWithTranslucency;
     ConstructorHelpers::FObjectFinder<UMaterialInstance> BaseMaterialWithWater;
     ConstructorHelpers::FObjectFinder<UTexture2D> Transparent1x1;
     FConstructorStatics()
         : BaseMaterial(TEXT(
               "/CesiumForUnreal/Materials/Instances/MI_CesiumThreeOverlaysAndClipping.MI_CesiumThreeOverlaysAndClipping")),
+          BaseMaterialWithTranslucency(TEXT(
+              "/CesiumForUnreal/Materials/Instances/MI_CesiumThreeOverlaysAndClippingTranslucent.MI_CesiumThreeOverlaysAndClippingTranslucent")),
           BaseMaterialWithWater(TEXT(
               "/CesiumForUnreal/Materials/Instances/MI_CesiumThreeOverlaysAndClippingAndWater.MI_CesiumThreeOverlaysAndClippingAndWater")),
           Transparent1x1(
@@ -2081,6 +2115,8 @@ UCesiumGltfComponent::UCesiumGltfComponent() : USceneComponent() {
   static FConstructorStatics ConstructorStatics;
 
   this->BaseMaterial = ConstructorStatics.BaseMaterial.Object;
+  this->BaseMaterialWithTranslucency =
+      ConstructorStatics.BaseMaterialWithTranslucency.Object;
   this->BaseMaterialWithWater = ConstructorStatics.BaseMaterialWithWater.Object;
   this->Transparent1x1 = ConstructorStatics.Transparent1x1.Object;