+++ Source/CesiumRuntime/Public/Cesium3DTileset.h
@@ -686,11 +686,12 @@ private:
   bool EnableWaterMask = false;
 
   /**
-   * A custom Material to use to render this tileset, in order to implement
-   * custom visual effects.
+   * A custom Material to use to render opaque elements in this tileset, in
+   * order to implement custom visual effects.
    *
-   * The custom material should generally be created by copying the
-   * "M_CesiumOverlay" material and customizing it as desired.
+   * The custom material should generally be created by copying the Material
+   * Instance "MI_CesiumThreeOverlaysAndClipping" and customizing the copy as
+   * desired.
    */
   UPROPERTY(
       EditAnywhere,
@@ -700,15 +701,30 @@ private:
   UMaterialInterface* Material = nullptr;
 
   /**
+   * A custom Material to use to render translucent elements of the tileset, in
+   * order to implement custom visual effects.
+   *
+   * The custom material should generally be created by copying the Material
+   * Instance "MI_CesiumThreeOverlaysAndClippingTranslucent" and customizing the
+   * copy as desired. Make sure that its Material Property Overrides -> Blend
+   * Mode is set to "Translucent".
+   */
+  UPROPERTY(
+      EditAnywhere,
+      BlueprintGetter = GetTranslucentMaterial,
+      BlueprintSetter = SetTranslucentMaterial,
+      Category = "Cesium|Rendering")
+  UMaterialInterface* TranslucentMaterial = nullptr;
+
+  /**
    * A custom Material to use to render this tileset in areas where the
    * watermask is, in order to implement custom visual effects.
    * Currently only applicable for quantized-mesh tilesets that support the
    * water mask extension.
    *
-   * The custom material should generally be created by copying the
-   * "M_CesiumOverlayWater" material and customizing it as desired. For best
-   * results, any changes to the above material should also be duplicated in the
-   * Water Material.
+   * The custom material should generally be created by copying the Material
+   * Instance "MI_CesiumThreeOverlaysAndClippingAndWater" and customizing the
+   * copy as desired.
    */
   UPROPERTY(
       EditAnywhere,
@@ -812,6 +828,14 @@ public:
 
   UFUNCTION(BlueprintSetter, Category = "Cesium|Rendering")
   void SetMaterial(UMaterialInterface* InMaterial);
+
+  UFUNCTION(BlueprintGetter, Category = "Cesium|Rendering")
+  UMaterialInterface* GetTranslucentMaterial() const {
+    return TranslucentMaterial;
+  }
+
+  UFUNCTION(BlueprintSetter, Category = "Cesium|Rendering")
+  void SetTranslucentMaterial(UMaterialInterface* InMaterial);
 
   UFUNCTION(BlueprintGetter, Category = "Cesium|Rendering")
   UMaterialInterface* GetWaterMaterial() const { return WaterMaterial; }